+++ radcad/__init__.py
@@ -3,3 +3,6 @@ __version__ = "0.0.0"
 from radcad.wrappers import Context, Model, Simulation, Experiment
 from radcad.engine import Engine
 from radcad.backends import Backend
+
+
+__pdoc__ = {"extensions": False}