+++ radcad/backends/single_process.py
@@ -6,6 +6,6 @@ class ExecutorSingleProcess(Executor):
     def execute_runs(self):
         result = [
             core._single_run_wrapper((config, self.engine.raise_exceptions))
-            for config in self.engine.run_generator
+            for config in self.engine._run_generator
         ]
         return result